[move] helpers.js moved out of core folder
