+++ lib/core/core.js
@@ -8,7 +8,7 @@
 
   var path = require('path');
 
-  var helpers = require('./helpers.js');
+  var helpers = require('../helpers.js');
   var Process = require(path.join(helpers.core, 'process.js'));
