+++ lib/core/monitorAbstract.js
@@ -9,7 +9,7 @@
   var util = require('util');
   var events = require('events');
 
-  var helpers = require('./helpers.js');
+  var helpers = require('../helpers.js');
 
   function MonitorAbstract(options, ready) {
     this.options = options;
