+++ lib/core/process.js
@@ -11,7 +11,7 @@
   var events = require('events');
   var child_process = require('child_process');
 
-  var helpers = require('./helpers.js');
+  var helpers = require('../helpers.js');
   var streams = require(path.join(helpers.core, 'streams.js'));
 
   // this constrcutor will spawn a new process using a option object
