+++ lib/executables/daemon.js
@@ -8,7 +8,7 @@
 
   var path = require('path');
 
-  var helpers = require('./helpers.js');
+  var helpers = require('../helpers.js');
   var streams = require(path.join(helpers.core, 'streams.js'));
   var core = require(path.join(helpers.core, 'core.js'));
