+++ lib/executables/execute.js
@@ -6,7 +6,7 @@
 (function () {
   "use strict";
 
-  var helpers = require('./helpers.js');
+  var helpers = require('../helpers.js');
   var path = require('path');
   var core = require(path.join(helpers.core, 'core.js'));
