+++ lib/executables/pump.js
@@ -6,7 +6,7 @@
 (function () {
   "use strict";
 
-  var helpers = require('./helpers.js');
+  var helpers = require('../helpers.js');
   var path = require('path');
 
   var streams = require(path.join(helpers.core, 'streams.js'));
