+++ lib/helpers.js
@@ -13,7 +13,7 @@
   exports.exists = fs.exists || path.exists;
 
   // common directorys
-  exports.root = path.join(path.dirname(module.filename), '../../');
+  exports.root = path.join(path.dirname(module.filename), '../');
   exports.lib = path.join(exports.root, 'lib');
   exports.core = path.join(exports.lib, 'core');
   exports.execute = path.join(exports.lib, 'executables');
